backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
cohere>=4.45
qdrant-client>=1.7.1
google-generativeai>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
httpx>=0.26.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
